+++ packages/core/src/workflows/types.ts
@@ -3,6 +3,13 @@ import type { z } from 'zod';
 import type { ExecuteFunction, Step } from './step';
 import type { SerializedStepFlowEntry } from './workflow';
 
+export type Emitter = {
+  emit: (event: string, data: any) => Promise<void>;
+  on: (event: string, callback: (data: any) => void) => void;
+  off: (event: string, callback: (data: any) => void) => void;
+  once: (event: string, callback: (data: any) => void) => void;
+};
+
 export type StepSuccess<P, R, S, T> = {
   status: 'success';
   output: T;
@@ -45,11 +52,20 @@ export type StepRunning<P, R, S> = {
   resumedAt?: number;
 };
 
+export type StepWaiting<P, R, S> = {
+  status: 'waiting';
+  payload: P;
+  suspendPayload?: S;
+  resumePayload?: R;
+  startedAt: number;
+};
+
 export type StepResult<P, R, S, T> =
   | StepSuccess<P, R, S, T>
   | StepFailure<P, R, S>
   | StepSuspended<P, S>
-  | StepRunning<P, R, S>;
+  | StepRunning<P, R, S>
+  | StepWaiting<P, R, S>;
 
 export type StepsRecord<T extends readonly Step<any, any, any>[]> = {
   [K in T[number]['id']]: Extract<T[number], { id: K }>;
@@ -60,17 +76,20 @@ export type DynamicMapping<TPrevSchema extends z.ZodTypeAny, TSchemaOut extends 
   schema: TSchemaOut;
 };
 
-export type PathsToStringProps<T> = T extends object
-  ? {
-      [K in keyof T]: T[K] extends object
-        ? K extends string
-          ? K | `${K}.${PathsToStringProps<T[K]>}`
-          : never
-        : K extends string
-          ? K
-          : never;
-    }[keyof T]
-  : never;
+export type PathsToStringProps<T> =
+  T extends z.ZodObject<infer V>
+    ? PathsToStringProps<V>
+    : T extends object
+      ? {
+          [K in keyof T]: T[K] extends object
+            ? K extends string
+              ? K | `${K}.${PathsToStringProps<T[K]>}`
+              : never
+            : K extends string
+              ? K
+              : never;
+        }[keyof T]
+      : never;
 
 export type ExtractSchemaType<T extends z.ZodType<any>> = T extends z.ZodObject<infer V> ? V : never;
 
@@ -98,6 +117,11 @@ export type StreamEvent =
       type: 'step-suspended';
       payload: any;
       id: string;
+    }
+  | {
+      type: 'step-waiting';
+      payload: any;
+      id: string;
     };
 
 export type WatchEvent = {
@@ -105,18 +129,18 @@ export type WatchEvent = {
   payload: {
     currentStep?: {
       id: string;
-      status: 'running' | 'success' | 'failed' | 'suspended';
+      status: 'running' | 'success' | 'failed' | 'suspended' | 'waiting';
       output?: Record<string, any>;
       resumePayload?: Record<string, any>;
       payload?: Record<string, any>;
       error?: string | Error;
     };
     workflowState: {
-      status: 'running' | 'success' | 'failed' | 'suspended';
+      status: 'running' | 'success' | 'failed' | 'suspended' | 'waiting';
       steps: Record<
         string,
         {
-          status: 'running' | 'success' | 'failed' | 'suspended';
+          status: 'running' | 'success' | 'failed' | 'suspended' | 'waiting';
           output?: Record<string, any>;
           payload?: Record<string, any>;
           resumePayload?: Record<string, any>;
@@ -152,7 +176,7 @@ export type ZodPathType<T extends z.ZodTypeAny, P extends string> =
 export interface WorkflowRunState {
   // Core state info
   runId: string;
-  status: 'success' | 'failed' | 'suspended' | 'running';
+  status: 'success' | 'failed' | 'suspended' | 'running' | 'waiting';
   result?: Record<string, any>;
   error?: string | Error;
   value: Record<string, string>;